firmware: dma_rtg OP_PAN — sign-extend x/y for framebuffer offset
data->x[0] / data->y[0] are uint16_t, so negative pan values wrapped
to huge positive offsets and produced garbage. Cast to int16_t before
the shift/multiply, matching the sprite_x_offset / sprite_y_offset
lines just above. Reported against upstream by the z3660 project.

+++ ZZ9000_proto.sdk/ZZ9000OS/src/dma_rtg.c
@@ -235,8 +235,8 @@ void handle_blitter_dma_op(struct ZZ_VIDEO_STATE* vs, uint16_t zdata)
 
             vs->framebuffer_pan_width = data->x[1];
             u32 framebuffer_color_format = data->u8_user[GFXDATA_U8_COLORMODE];
-            vs->framebuffer_pan_offset = data->offset[0] + (data->x[0] << data->u8_user[GFXDATA_U8_COLORMODE]);
-            vs->framebuffer_pan_offset += (data->y[0] * (vs->framebuffer_pan_width << framebuffer_color_format));
+            vs->framebuffer_pan_offset = data->offset[0] + (((int16_t)data->x[0]) << data->u8_user[GFXDATA_U8_COLORMODE]);
+            vs->framebuffer_pan_offset += (((int16_t)data->y[0]) * (vs->framebuffer_pan_width << framebuffer_color_format));
             break;
 
         case OP_SET_SPLIT_POS:
